Updated to latest version of TheDistanceKit and set up for Jenkins.

+++ NHS Prase/Model/Persistent Model/Machine/_PRAnswerOption.h
@@ -41,31 +41,37 @@ extern const struct PRAnswerOptionFetchedProperties {
 
 
 
-@property (nonatomic, strong) NSNumber *answerID;
+
+@property (nonatomic, strong) NSNumber* answerID;
 
 
-@property long long answerIDValue;
-- (long long)answerIDValue;
-- (void)setAnswerIDValue:(long long)value_;
+
+@property int64_t answerIDValue;
+- (int64_t)answerIDValue;
+- (void)setAnswerIDValue:(int64_t)value_;
 
 //- (BOOL)validateAnswerID:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) NSNumber *imageID;
+
+@property (nonatomic, strong) NSNumber* imageID;
 
 
-@property long long imageIDValue;
-- (long long)imageIDValue;
-- (void)setImageIDValue:(long long)value_;
+
+@property int64_t imageIDValue;
+- (int64_t)imageIDValue;
+- (void)setImageIDValue:(int64_t)value_;
 
 //- (BOOL)validateImageID:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) NSString *imageName;
+
+@property (nonatomic, strong) NSString* imageName;
+
 
 
 //- (BOOL)validateImageName:(id*)value_ error:(NSError**)error_;
@@ -73,7 +79,9 @@ extern const struct PRAnswerOptionFetchedProperties {
 
 
 
-@property (nonatomic, strong) NSString *imageTintIdentifier;
+
+@property (nonatomic, strong) NSString* imageTintIdentifier;
+
 
 
 //- (BOOL)validateImageTintIdentifier:(id*)value_ error:(NSError**)error_;
@@ -81,19 +89,23 @@ extern const struct PRAnswerOptionFetchedProperties {
 
 
 
-@property (nonatomic, strong) NSNumber *localizationID;
+
+@property (nonatomic, strong) NSNumber* localizationID;
 
 
-@property long long localizationIDValue;
-- (long long)localizationIDValue;
-- (void)setLocalizationIDValue:(long long)value_;
+
+@property int64_t localizationIDValue;
+- (int64_t)localizationIDValue;
+- (void)setLocalizationIDValue:(int64_t)value_;
 
 //- (BOOL)validateLocalizationID:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) NSString *title;
+
+@property (nonatomic, strong) NSString* title;
+
 
 
 //- (BOOL)validateTitle:(id*)value_ error:(NSError**)error_;
@@ -102,9 +114,10 @@ extern const struct PRAnswerOptionFetchedProperties {
 
 
 
-@property (nonatomic, strong) NSSet* answerSet;
+@property (nonatomic, strong) NSSet *answerSet;
 
 - (NSMutableSet*)answerSetSet;
+
 
 
 
@@ -126,8 +139,8 @@ extern const struct PRAnswerOptionFetchedProperties {
 - (NSNumber*)primitiveAnswerID;
 - (void)setPrimitiveAnswerID:(NSNumber*)value;
 
-- (long long)primitiveAnswerIDValue;
-- (void)setPrimitiveAnswerIDValue:(long long)value_;
+- (int64_t)primitiveAnswerIDValue;
+- (void)setPrimitiveAnswerIDValue:(int64_t)value_;
 
 
 
@@ -135,8 +148,8 @@ extern const struct PRAnswerOptionFetchedProperties {
 - (NSNumber*)primitiveImageID;
 - (void)setPrimitiveImageID:(NSNumber*)value;
 
-- (long long)primitiveImageIDValue;
-- (void)setPrimitiveImageIDValue:(long long)value_;
+- (int64_t)primitiveImageIDValue;
+- (void)setPrimitiveImageIDValue:(int64_t)value_;
 
 
 
@@ -156,8 +169,8 @@ extern const struct PRAnswerOptionFetchedProperties {
 - (NSNumber*)primitiveLocalizationID;
 - (void)setPrimitiveLocalizationID:(NSNumber*)value;
 
-- (long long)primitiveLocalizationIDValue;
-- (void)setPrimitiveLocalizationIDValue:(long long)value_;
+- (int64_t)primitiveLocalizationIDValue;
+- (void)setPrimitiveLocalizationIDValue:(int64_t)value_;
