+++ NHS Prase/Model/Persistent Model/Machine/_PRHospital.h
@@ -35,19 +35,23 @@ extern const struct PRHospitalFetchedProperties {
 
 
 
-@property (nonatomic, strong) NSNumber *id;
+
+@property (nonatomic, strong) NSNumber* id;
 
 
-@property long long idValue;
-- (long long)idValue;
-- (void)setIdValue:(long long)value_;
+
+@property int64_t idValue;
+- (int64_t)idValue;
+- (void)setIdValue:(int64_t)value_;
 
 //- (BOOL)validateId:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) NSString *name;
+
+@property (nonatomic, strong) NSString* name;
+
 
 
 //- (BOOL)validateName:(id*)value_ error:(NSError**)error_;
@@ -56,16 +60,17 @@ extern const struct PRHospitalFetchedProperties {
 
 
 
-@property (nonatomic, strong) PRTrust* trust;
+@property (nonatomic, strong) PRTrust *trust;
 
 //- (BOOL)validateTrust:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) NSSet* wards;
+@property (nonatomic, strong) NSSet *wards;
 
 - (NSMutableSet*)wardsSet;
+
 
 
 
@@ -87,8 +92,8 @@ extern const struct PRHospitalFetchedProperties {
 - (NSNumber*)primitiveId;
 - (void)setPrimitiveId:(NSNumber*)value;
 
-- (long long)primitiveIdValue;
-- (void)setPrimitiveIdValue:(long long)value_;
+- (int64_t)primitiveIdValue;
+- (void)setPrimitiveIdValue:(int64_t)value_;
